% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpression_analysis.R
\name{calculate_coexpression_stats}
\alias{calculate_coexpression_stats}
\title{Calculate co-expression statistics}
\usage{
calculate_coexpression_stats(cor_result, include_pairwise = FALSE)
}
\arguments{
\item{cor_result}{Result from calculate_isoform_coexpression}

\item{include_pairwise}{Include pairwise correlation details}
}
\value{
A list of statistical summaries
}
\description{
Generates comprehensive statistics for co-expression analysis results.
}
