% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_metrics.R
\name{.calculate_cell_type_coexpression_variability}
\alias{.calculate_cell_type_coexpression_variability}
\title{Calculate CV of cell-type-specific co-expression varaibility}
\usage{
.calculate_cell_type_coexpression_variability(scht_obj, gene)
}
\arguments{
\item{scht_obj}{Single-Cell Hierarchical Tensor object}

\item{gene}{Gene name to analyse}
}
\value{
List containing CV and cell type-specific mean IDI values
}
\description{
This function calculates the coefficient of variation of mean intra-cellular
diversity across cell types. High values indicate that a gene employs dramatically
different isoform co-expression patterns in different cell types.
}
\keyword{internal}
