% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_metrics.R
\name{.calculate_js_distance}
\alias{.calculate_js_distance}
\title{Calculate Jensen-Shannon distance between two probability distributions}
\usage{
.calculate_js_distance(p, q)
}
\arguments{
\item{p}{First probability vector}

\item{q}{Second probability vector}
}
\value{
Jensen-Shannon distance
}
\description{
Calculates Jensen-Shannon distance between two probability vectors,
taking advantage of efficiency by operating only on non-zero elements.
}
\keyword{internal}
