% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scht.R
\name{.create_integrated_scht}
\alias{.create_integrated_scht}
\title{Create integrated SCHT}
\usage{
.create_integrated_scht(scht_obj, cell_type_result)
}
\arguments{
\item{scht_obj}{Original SCHT object}

\item{cell_type_result}{cell type-specific analysis results}
}
\value{
An IntegratedSCHT object
}
\description{
Creates an integrated data structure combining original SCHT and
cell type-specific analyses with efficient memory usage.
}
\keyword{internal}
