% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexity_metrics.R
\name{.find_optimal_lambda}
\alias{.find_optimal_lambda}
\title{Determine optimal Yeo-Johnson transformation parameter}
\usage{
.find_optimal_lambda(values)
}
\arguments{
\item{values}{Numeric vector to transform}
}
\value{
Optimal lambda parameter
}
\description{
Determine optimal Yeo-Johnson transformation parameter
}
\keyword{internal}
