% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_analysis.R
\name{.get_colour_palette}
\alias{.get_colour_palette}
\title{Get color palette with fallback options}
\usage{
.get_colour_palette(n, palette_name = "Set3")
}
\arguments{
\item{n}{Number of colors needed}

\item{palette_name}{Name of the RColorBrewer palette}
}
\value{
Vector of colors
}
\description{
Internal function to get color palettes with fallbacks when RColorBrewer is not available
}
\keyword{internal}
