% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_utils.R
\name{.get_fallback_colours}
\alias{.get_fallback_colours}
\title{Get alternative colour palette}
\usage{
.get_fallback_colours(n, type = "qualitative")
}
\arguments{
\item{n}{Number of colours needed}

\item{type}{Type of palette ("sequential", "diverging", "qualitative")}
}
\value{
Vector of colour codes
}
\description{
Internal function to provide fallback colour palettes when packages are not available
}
\keyword{internal}
