% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_analysis.R
\name{.get_tc_palette}
\alias{.get_tc_palette}
\title{Get colour palette for transcriptome complexity visualisation}
\usage{
.get_tc_palette()
}
\value{
A list with colour specifications
}
\description{
Internal function to create a consistent colour palette for visualisations.
}
\keyword{internal}
