% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scht.R
\name{.perform_log_transform}
\alias{.perform_log_transform}
\title{Perform log transformation}
\usage{
.perform_log_transform(gene_counts_norm, transcript_counts_norm, verbose)
}
\arguments{
\item{gene_counts_norm}{Normalised gene counts}

\item{transcript_counts_norm}{Normalised transcript counts}
}
\value{
A list containing transformed data
}
\description{
Applies log2(x+1) transformation to normalised count data with
special handling for different matrix types.
}
\keyword{internal}
