% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_scht.R
\name{.perform_normalisation}
\alias{.perform_normalisation}
\title{Perform normalisation}
\usage{
.perform_normalisation(gene_counts_filtered, transcript_counts_filtered)
}
\arguments{
\item{gene_counts_filtered}{Filtered gene counts}

\item{transcript_counts_filtered}{Filtered transcript counts}
}
\value{
A list containing normalised data
}
\description{
Normalises count data to counts per million (CPM) using
operations to maintain memory efficiency.
}
\keyword{internal}
