% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation_analysis.R
\name{.tc_theme}
\alias{.tc_theme}
\title{Theme for transcriptomic complexity visualisation}
\usage{
.tc_theme()
}
\value{
A ggplot2 theme object
}
\description{
Internal function to create a consistent theme for plots.
}
\keyword{internal}
