% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_counts_blood}
\alias{gene_counts_blood}
\title{Blood dataset gene expression counts}
\format{
A data frame
}
\source{
Wang et al. (2022). Science Advances 8(1):eabg5369. Data from https://zenodo.org/records/5706781. Licensed under CC BY 4.0.
}
\description{
Gene-level expression count matrix from blood single-cell RNA-seq data.
This data is derived from Wang et al. (2022) study on alternative splicing
during hematopoietic stem cell formation.
}
\references{
Wang F, et al. (2022). Single-cell architecture and functional requirement of alternative splicing during hematopoietic stem cell formation. Science Advances 8(1):eabg5369.
}
\keyword{datasets}
