% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample2stage}
\alias{sample2stage}
\title{Blood dataset cell metadata}
\format{
A data frame with columns:
\describe{
\item{sample}{Sample identifier}
\item{cell_type}{Cell type annotation}
}
}
\source{
Wang et al. (2022). Science Advances 8(1):eabg5369. Data from https://zenodo.org/records/5706781. Licensed under CC BY 4.0.
}
\description{
Cell metadata including cell type annotations for the blood dataset.
}
\references{
Wang F, et al. (2022). Single-cell architecture and functional requirement of alternative splicing during hematopoietic stem cell formation. Science Advances 8(1):eabg5369.
}
\keyword{datasets}
