% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcript_info}
\alias{transcript_info}
\title{Transcript information}
\format{
A data frame with columns:
\describe{
\item{transcript_id}{Transcript identifier}
\item{transcript_name}{Transcript name}
\item{gene_id}{Gene identifier}
\item{gene_name}{Gene symbol}
\item{transcript_type}{Transcript biotype}
\item{gene_type}{Gene biotype}
}
}
\source{
GENCODE Mouse Release M36 from https://www.gencodegenes.org/mouse/release_M36.html. Processed using create_transcript_info() function from ScIsoX package.
}
\description{
Transcript and gene annotation information from mouse genome.
This data was processed using the create_transcript_info() function from ScIsoX package.
}
\seealso{
\code{\link{create_transcript_info}} for the function used to generate this data
}
\keyword{datasets}
